requirements.txt
flask
flask-cors
pymongo[srv]
dnspython
twilio
openai
python-dotenv
pytz
certifi
waitress
apscheduler
